fixing workflow - moved get_and_upload function
